fix wrapped component test
