+++ test/react-web/client/graphql/shared-operations-1.test.tsx
@@ -16,7 +16,7 @@ import {
 import graphql, { withApollo } from '../../../../src/graphql';
 import { compose } from '../../../../src/';
 
-describe('shared opertations', () => {
+describe('shared operations', () => {
 
   describe('withApollo', () => {
     it('passes apollo-client to props', () => {
@@ -134,40 +134,7 @@ describe('shared opertations', () => {
     (wrapper as any).unmount();
   });
 
-  // XXX fix test
-  xit('allows a way to access the wrapped component instance', () => {
-    const query = gql`query people { allPeople(first: 1) { people { name } } }`;
-    const data = { allPeople: { people: [ { name: 'Luke Skywalker' } ] } };
-    const networkInterface = mockNetworkInterface({ request: { query }, result: { data } });
-    const client = new ApolloClient({ networkInterface });
-
-    const testData = { foo: 'bar' };
-
-    class Container extends React.Component<any, any> {
-      someMethod() {
-        return testData;
-      }
-
-      render() {
-        return <span></span>;
-      }
-    }
-
-    const Decorated = graphql(query, { withRef: true })(Container);
-
-    /* const tree = TestUtils.renderIntoDocument( */
-    /*   <ProviderMock client={client}> */
-    /*     <Decorated /> */
-    /*   </ProviderMock> */
-    /* ) as any; */
-
-    /* const decorated = TestUtils.findRenderedComponentWithType(tree, Decorated); */
-
-    /* expect(() => (decorated as any).someMethod()).toThrow(); */
-    /* expect((decorated as any).getWrappedInstance().someMethod()).toEqual(testData); */
-    /* expect((decorated as any).refs.wrappedInstance.someMethod()).toEqual(testData); */
-
-  });
+  // XXX move shared-operations-2.tsx back here when React 15.4 is released
 
   it('allows options to take an object', (done) => {
     const query = gql`query people { allPeople(first: 1) { people { name } } }`;
